refactor: rename components to ensure backwards compatibility.
