+++ packages/map-template/src/components/MapTemplate/MapTemplate.scss
